dto 도메인별로 정리
