--- src/main/java/com/sprint/mission/discodeit/dto/FindBinaryContentRequestDto.java
@@ -1,11 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-import java.time.Instant;
-
-public record FindBinaryContentRequestDto(
-        String fileName,
-        String contentType,
-        byte[] bytes,
-        Instant createdAt
-) {
-}
