--- src/main/java/com/sprint/mission/discodeit/dto/FindMessageByChannelIdResponseDto.java
@@ -1,15 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-import java.time.Instant;
-import java.util.List;
-import java.util.UUID;
-
-public record FindMessageByChannelIdResponseDto(
-        UUID messageId,
-        String nickname,
-        List<UUID> attachmentList,
-        String content,
-        Instant createAt
-) {
-
-}
