--- src/main/java/com/sprint/mission/discodeit/dto/LoginRequest.java
@@ -1,7 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-public record LoginRequest(
-        String username,
-        String password
-) {
-}
