--- src/main/java/com/sprint/mission/discodeit/dto/MessageCreateRequest.java
@@ -1,10 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-import java.util.UUID;
-
-public record MessageCreateRequest(
-        UUID channelId,
-        UUID userId,
-        String content
-) {
-}
