--- src/main/java/com/sprint/mission/discodeit/dto/PublicChannelCreateRequest.java
@@ -1,10 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-import java.util.List;
-import java.util.UUID;
-
-public record PublicChannelCreateRequest(
-    String channelName
-) {
-
-}
