--- src/main/java/com/sprint/mission/discodeit/dto/SaveUserStatusParamDto.java
@@ -1,8 +1,0 @@
-package com.sprint.mission.discodeit.dto;
-
-import java.util.UUID;
-
-public record SaveUserStatusParamDto(
-        UUID userUUID
-) {
-}
