+++ src/main/java/com/sprint/mission/discodeit/dto/auth/LoginRequest.java
@@ -1,0 +1,8 @@
+package com.sprint.mission.discodeit.dto.auth;
+
+public record LoginRequest(
+    String username,
+    String password
+) {
+
+}
