+++ src/main/java/com/sprint/mission/discodeit/dto/binaryContent/FindBinaryContentRequestDto.java
@@ -1,0 +1,12 @@
+package com.sprint.mission.discodeit.dto.binaryContent;
+
+import java.time.Instant;
+
+public record FindBinaryContentRequestDto(
+    String fileName,
+    String contentType,
+    byte[] bytes,
+    Instant createdAt
+) {
+
+}
