+++ src/main/java/com/sprint/mission/discodeit/dto/binaryContent/SaveBinaryContentRequestDto.java
@@ -1,6 +1,5 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.binaryContent;
 
-import io.swagger.v3.oas.annotations.media.Schema;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
