+++ src/main/java/com/sprint/mission/discodeit/dto/channel/PublicChannelCreateRequest.java
@@ -1,0 +1,7 @@
+package com.sprint.mission.discodeit.dto.channel;
+
+public record PublicChannelCreateRequest(
+    String channelName
+) {
+
+}
