+++ src/main/java/com/sprint/mission/discodeit/dto/channel/SaveChannelResponseDto.java
@@ -1,0 +1,15 @@
+package com.sprint.mission.discodeit.dto.channel;
+
+import com.sprint.mission.discodeit.constant.ChannelType;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record SaveChannelResponseDto(
+    UUID channelUUID,
+    String channelName,
+    ChannelType channelType,
+    Instant createdAt
+) {
+
+}
