+++ src/main/java/com/sprint/mission/discodeit/dto/message/FindMessageByChannelIdResponseDto.java
@@ -1,0 +1,15 @@
+package com.sprint.mission.discodeit.dto.message;
+
+import java.time.Instant;
+import java.util.List;
+import java.util.UUID;
+
+public record FindMessageByChannelIdResponseDto(
+    UUID messageId,
+    String nickname,
+    List<UUID> attachmentList,
+    String content,
+    Instant createAt
+) {
+
+}
