+++ src/main/java/com/sprint/mission/discodeit/dto/message/MessageCreateRequest.java
@@ -1,0 +1,11 @@
+package com.sprint.mission.discodeit.dto.message;
+
+import java.util.UUID;
+
+public record MessageCreateRequest(
+    UUID channelId,
+    UUID userId,
+    String content
+) {
+
+}
