+++ src/main/java/com/sprint/mission/discodeit/dto/message/MessageUpdateRequest.java
@@ -1,6 +1,7 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.message;
 
 public record MessageUpdateRequest(
     String content
 ) {
+
 }
