+++ src/main/java/com/sprint/mission/discodeit/dto/readStatus/FindIsReadStatusDto.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.readStatus;
 
 import java.util.UUID;
 
@@ -6,4 +6,5 @@ public record FindIsReadStatusDto(
     UUID channelUUID,
     String channelName
 ) {
+
 }
