+++ src/main/java/com/sprint/mission/discodeit/dto/readStatus/ReadStatusRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.readStatus;
 
 import java.time.Instant;
 import java.util.UUID;
