+++ src/main/java/com/sprint/mission/discodeit/dto/user/FindUserDto.java
@@ -1,0 +1,17 @@
+package com.sprint.mission.discodeit.dto.user;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record FindUserDto(
+    UUID userUUID,
+    String username,
+    String email,
+    UUID profileId,
+    Instant createdAt,
+    Instant updatedAt,
+    Instant lastLoginTime,
+    boolean online
+) {
+
+}
