+++ src/main/java/com/sprint/mission/discodeit/dto/user/SaveUserRequestDto.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.user;
 
 public record SaveUserRequestDto(
     String username,
