+++ src/main/java/com/sprint/mission/discodeit/dto/userStatus/SaveUserStatusParamDto.java
@@ -1,0 +1,9 @@
+package com.sprint.mission.discodeit.dto.userStatus;
+
+import java.util.UUID;
+
+public record SaveUserStatusParamDto(
+    UUID userUUID
+) {
+
+}
