+++ src/main/java/com/sprint/mission/discodeit/dto/userStatus/UserStatusUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.userStatus;
 
 import java.time.Instant;
