+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.dto.LoginRequest;
+import com.sprint.mission.discodeit.dto.auth.LoginRequest;
 
 public interface AuthService {
